backend/requirements.txt
flask
flask-cors
torch
transformers
pillow
opencv-python-headless
numpy
gunicorn